Remove `ngraph::helper` (#22704)
### Details:
 - Port required functions to `common_test_utils/test_helpers.*pp`
 - Delete `ov_models` target

### Tickets:
 - [CVS-123998](https://jira.devtools.intel.com/browse/CVS-123998)
