+++ src/plugins/hetero/tests/unit/CMakeLists.txt
@@ -21,7 +21,6 @@ ov_add_test_target(
             ${OBJ_LIB}
         LINK_LIBRARIES
             unit_test_utils
-            ov_models
         DEPENDENCIES
             mock_engine
         ADD_CLANG_FORMAT
