+++ src/plugins/intel_cpu/tests/functional/custom/single_layer_tests/non_max_suppression.cpp
@@ -8,6 +8,7 @@
 #include "shared_test_classes/base/ov_subgraph.hpp"
 #include "shared_test_classes/base/utils/ranges.hpp"
 #include "utils/cpu_test_utils.hpp"
+#include "common_test_utils/test_enums.hpp"
 
 using namespace CPUTestUtils;
