+++ src/plugins/intel_cpu/tests/functional/custom/single_layer_tests/pad.cpp
@@ -8,6 +8,7 @@
 #include "openvino/op/pad.hpp"
 #include "shared_test_classes/base/ov_subgraph.hpp"
 #include "utils/cpu_test_utils.hpp"
+#include "common_test_utils/test_enums.hpp"
 
 using namespace CPUTestUtils;
