+++ src/plugins/intel_cpu/tests/functional/custom/single_layer_tests/variadic_split.cpp
@@ -6,6 +6,7 @@
 #include "common_test_utils/test_enums.hpp"
 #include "shared_test_classes/base/ov_subgraph.hpp"
 #include "utils/cpu_test_utils.hpp"
+#include "common_test_utils/test_enums.hpp"
 
 using namespace ov::test;
 using namespace CPUTestUtils;
