+++ src/plugins/intel_cpu/tests/functional/custom/subgraph_tests/src/denormal_check.cpp
@@ -3,7 +3,7 @@
 //
 
 #include "shared_test_classes/base/ov_subgraph.hpp"
-#include "ov_models/utils/data_utils.hpp"
+#include "common_test_utils/data_utils.hpp"
 #include "openvino/runtime/aligned_buffer.hpp"
 
 namespace ov {
