+++ src/plugins/intel_gpu/tests/functional/single_layer_tests/dynamic/dft.cpp
@@ -5,7 +5,6 @@
 #include "shared_test_classes/base/ov_subgraph.hpp"
 #include "common_test_utils/ov_tensor_utils.hpp"
 #include "common_test_utils/test_enums.hpp"
-#include "ov_models/utils/ov_helpers.hpp"
 
 #include "openvino/op/parameter.hpp"
 #include "openvino/op/constant.hpp"
