+++ src/tests/functional/plugin/shared/src/low_precision_transformations/convolution_transformation.cpp
@@ -11,7 +11,6 @@
 
 
 #include "common_test_utils/common_utils.hpp"
-#include "ov_models/pass/convert_prc.hpp"
 #include "ov_lpt_models/fake_quantize_and_convolution.hpp"
 
 namespace LayerTestsDefinitions {
