+++ src/tests/functional/plugin/shared/src/low_precision_transformations/fake_quantize_and_two_output_branches_with_convolution.cpp
@@ -11,7 +11,6 @@
 
 
 #include "common_test_utils/common_utils.hpp"
-#include "ov_models/pass/convert_prc.hpp"
 
 namespace LayerTestsDefinitions {
