+++ src/tests/functional/plugin/shared/src/low_precision_transformations/normalize_transformation.cpp
@@ -12,7 +12,6 @@
 
 #include "common_test_utils/common_utils.hpp"
 
-#include "ov_models/pass/convert_prc.hpp"
 #include "ov_lpt_models/normalize_l2.hpp"
 
 namespace LayerTestsDefinitions {
