+++ src/tests/functional/plugin/shared/src/low_precision_transformations/output_layers_concat_multi_channel.cpp
@@ -13,7 +13,6 @@
 #include "common_test_utils/common_utils.hpp"
 #include "common_test_utils/node_builders/fake_quantize.hpp"
 
-#include "ov_models/pass/convert_prc.hpp"
 
 namespace LayerTestsDefinitions {
