+++ src/tests/functional/shared_test_classes/src/base/low_precision_transformations/layer_transformation.cpp
@@ -7,7 +7,6 @@
 #include <vector>
 #include <string>
 
-#include "ov_models/pass/convert_prc.hpp"
 
 namespace LayerTestsUtils {
 ov::pass::low_precision::LayerTransformation::Params LayerTransformationParamsNGraphFactory::createParamsU8I8AndI8() {
