+++ src/tests/ov_helpers/CMakeLists.txt
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-add_subdirectory(ov_models)
+add_subdirectory(ov_models) # TODO (vurusovs): delete after migration of NVidia and NPU
 add_subdirectory(ov_lpt_models)
 
 if(TARGET openvino::snippets)
